🔄 move CommonVals to index
