+++ Framework/Index/CommonVals.js
