+++ Framework/UI/Dialogs/Dialog.js
@@ -3,7 +3,7 @@
 *****************************************************************************************/
 import React, { memo } from 'react';
 import { View } from 'react-native';
-import { padSize, padSize2, padSize4 } from '../../CommonVals';
+import { padSize, padSize2, padSize4 } from '../../Index/CommonVals';
 import { Modal, Button, Card, Text, useTheme } from 'react-native-paper';
 
 /**
