+++ projects/TemplateApp/Screens/SampleContainersScreen.js
@@ -6,7 +6,7 @@ import { View } from 'react-native';
 // UI
 import { useTheme, Text } from 'react-native-paper';
 import { LinearLayout, ScreenLayout } from '../../../Framework/Index/UI';
-import { padSize } from '../../../Framework/CommonVals';
+import { padSize } from '../../../Framework/Index/CommonVals';
 
 /**
  * containers demos screen
