+++ projects/TemplateApp/Screens/SampleLayoutsScreen.js
@@ -5,7 +5,7 @@ import React from 'react';
 import { View } from 'react-native';
 import { useTheme, Text, Appbar } from 'react-native-paper';
 import { LinearLayout, GridLayout, ScreenLayout } from '../../../Framework/Index/UI';
-import { padSize } from '../../../Framework/CommonVals';
+import { padSize } from '../../../Framework/Index/CommonVals';
 
 /**
  * SampleLayoutsScreen Component
